% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broad_crested_wier.r
\name{broad_crested_weir}
\alias{broad_crested_weir}
\title{Broad crested weir equation}
\usage{
broad_crested_weir(w = NULL, h = NULL)
}
\arguments{
\item{w}{width of the weir in meters}

\item{h}{head (difference in water surface elevation and weir elevation) in meaters}
}
\description{
This function simulates individual capture histories for PSPAP type monitoring program for the Missouri River
}
\examples{
weir()
}
\keyword{equation}
\keyword{weir}
