% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_off.r
\name{In_out_el}
\alias{In_out_el}
\title{On off function}
\usage{
In_out_el(location, WSE, discharge)
}
\arguments{
\item{location}{of inflow/outflow to lake (Bridge_1, Bridge_2, Notch_1, WCS2, WCS2, Notch_2)}

\item{WSE}{water surface elevation in meters above sea level}

\item{discharge}{in cubic meters per second}
}
\value{
x = indictor of water going into the lake (1), no connection (0), out of lake (-1)
}
\description{
This function allows water to enter, leave, or stay in bluff lake depending on location and lake elevation.
}
\examples{
weir()
}
\keyword{equation}
\keyword{weir}
