% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharp_crested_wier.r
\name{sharp_crested_weir}
\alias{sharp_crested_weir}
\title{Weir equation function https://www.engineeringtoolbox.com/weirs-flow-rate-d_592.html}
\usage{
sharp_crested_weir(w = NULL, h = NULL)
}
\arguments{
\item{w}{width of the weir in meters}

\item{h}{head (difference in water surface elevation and weir elevation) in meaters}
}
\description{
This function simulates individual capture histories for PSPAP type monitoring program for the Missouri River
}
\examples{
weir()
}
\keyword{equation}
\keyword{weir}
