% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wse_sim.r
\name{wse_sim}
\alias{wse_sim}
\title{lake hydrodynanamic model}
\usage{
wse_sim(t, x, parms)
}
\arguments{
\item{t}{what time step}

\item{x}{lake volume in cubic meters}

\item{parms}{a vector of parameters used to simulate lake dynamics}
}
\value{
a list of lake volume, water surface elvation, lake elevation (same as water surface elevation the amount of water entering the lake, amount of water leaving the lake through the emergency spillway, and the amount of water going out the water control structure.
}
\description{
This function simulates water surface elevation dynamics for Bluff Lake given inflows and outflows.
}
\examples{
weir()
}
\keyword{equation}
\keyword{weir}
